Refactoring
- Moving special methods to ReassingTransformation
